Recast pid type as a u32
`pid_t` from Linux is allowed to be negative only for the purposes of
returning errors (for example `fork` failing). Actual running PIDs are
always positive.

We could alternatively use `std::num::NonZeroU32` but that type is very
clunky to work with: you can use `as i32` casts or compare against
`u32`s like the `PID_LIMIT` without extra calls (`Into::into` and
`NonZeroU32::get`, respectively). We don't need to the non-zero aspect
of the PID anywhere so I opted for a `u32` instead.