windows: Determine file handle object type dynamically
Handle object type IDs appear to be unstable across Windows versions.
On Windows 10 I see files as ID 37 while on Windows 11 they are 40.
The values are undocumented as far as I know so instead of constant
values, we should dynamically determine the correct ID number. We can
use the [NtQueryObject] system call with the `OBJECT_TYPE_INFORMATION`
object information class to get a string for the name of the type to
which a given handle points. There is a restriction though that the
call will only succeed if it runs against handles belonging to the
current process. So we first find the type ID by finding a file handle
owned by the current process. Then we count the handles with this ID
belonging to the queried process.

[NtQueryObject]: https://learn.microsoft.com/en-us/windows/win32/api/winternl/nf-winternl-ntqueryobject